IMPORTANT pGina code changes
1. Some broken code removed
Unlocking the locked user session is impossible in Windows 10 in some cases because of Microsoft changes
MutonUfoAi is aware of this problem but hasn't found a solution yet.
2. Locked users are now authenticated by plugins.
How is the user with a smart card supposed to log in via password?
I know that my solution is not perfect but I need it.

Minor MFLogin plugin improvements:
SQL query moved from file to a string constant
Default local database path set to pGina folder in Program Files